requirements.txt
streamlit
google-generativeai
edge-tts
pymupdf
pillow